Module renames for readability
