+++ phylo/src/substitution_models/protein_models/mod.rs
@@ -7,17 +7,17 @@ use crate::substitution_models::{
 };
 use crate::Result;
 
-pub(crate) mod protein_model_generics;
-pub(crate) use protein_model_generics::*;
+pub(crate) mod protein_generics;
+pub(crate) use protein_generics::*;
+pub(crate) mod protein_subst_params;
+pub(crate) use protein_subst_params::*;
+
 pub(crate) type ProteinSubstArray = [f64; 400];
 pub(crate) type ProteinFrequencyArray = [f64; 20];
 
 pub type ProteinSubstModel = SubstModel<ProteinSubstParams>;
 pub type ProteinSubstModelInfo = SubstModelInfo<ProteinSubstModel>;
 pub type ProteinLikelihoodCost<'a> = SubstLikelihoodCost<'a, ProteinSubstModel>;
-
-mod protein_subst_params;
-pub use protein_subst_params::*;
 
 impl SubstitutionModel for ProteinSubstModel {
     type ModelType = ProteinModelType;
